flash: handle TiCI estimate count RPC (#10826)
flash: handle GetEstimateTiCICount RPC for TiCI fulltext stats

(cherry picked from commit 4d9b25a7c5c8542acbd1b48c486ffff98934476b)
