+++ dbms/src/Flash/EstimateTiCICountHandler.h
@@ -1,0 +1,40 @@
+// Copyright 2026 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <Core/Types.h>
+#include <common/logger_useful.h>
+#include <grpcpp/support/status.h>
+#include <kvproto/coprocessor.pb.h>
+
+namespace DB
+{
+class EstimateTiCICountHandler
+{
+public:
+    EstimateTiCICountHandler(
+        const coprocessor::TiCIEstimateCountRequest * request_,
+        coprocessor::TiCIEstimateCountResponse * response_,
+        const String & identifier);
+
+    grpc::Status execute();
+
+private:
+    const coprocessor::TiCIEstimateCountRequest * request;
+    coprocessor::TiCIEstimateCountResponse * response;
+    const LoggerPtr log;
+};
+
+} // namespace DB
