+++ dbms/src/Flash/FlashService.h
@@ -100,6 +100,10 @@ public:
         const kvrpcpb::CompactRequest * request,
         kvrpcpb::CompactResponse * response) override;
 
+    grpc::Status GetEstimateTiCICount(
+        grpc::ServerContext * grpc_context,
+        const coprocessor::TiCIEstimateCountRequest * request,
+        coprocessor::TiCIEstimateCountResponse * response) override;
 
     // For S3 Lock Service
     grpc::Status tryAddLock(
